git-svn-id: svn+ssh://elmi.iti.uni-luebeck.de/export/svn/Hanse/branches/threadv1@938 8f13a6d9-6203-4c43-9b9f-a3039379152f

+++ Hanse/Behaviour_WallFollowing/wallfollowingform.cpp
@@ -67,7 +67,7 @@ void WallFollowingForm::on_startButton_clicked()
     wallfollow->setSettingsValue("corridorWidth", ui->corridorInput->text());
     wallfollow->setSettingsValue("wallTimer", ui->timerInput->text());
     //qDebug("startButton clicked");
-    wallfollow->setSettingsValue("initHeading", ui->headingInput->text());
+    wallfollow->setSettingsValue("initHeading", ui->headingInput->text().toFloat());
     wallfollow->setSettingsValue("useInitHeading", ui->headingBox->isChecked());
     QTimer::singleShot(0,wallfollow,SLOT(reset()));
     //qDebug("startButton clicked2");
